12869 invalidate payment attempts awaiting reader
It’s not possible to cancel reader connection.

The cancellation handler we’re adding is to cancel the `collectPayment` task – if the reader connection succeeds, that’s fine.

We need to prevent the payment from going ahead, especially so that PaymentIntents aren’t created, or if they are, they should be cancelled.

This commit adds an `invalidate` flag which is checked before we go on with the payment. If it’s true, the payment doesn’t start.