12869 invalidate payment orchestrator when cancelling
If we need to cancel a `collectPayment` task before we’ve got to the point of making the PaymentIntent, the `CardPresentPaymentStore.cancelPayment` function won’t do anything, because it doesn’t have anything to cancel.

This can happen if the task gets cancelled in the middle of connecting a reader, or while checking that the order is valid.

Instead, we want to prevent the payment intent from being created in the first place. That’s the responsibility of the PaymentCaptureOrchestrator, a dependency of the CollectOrderCardPaymentUseCase we’re adapting here.

By passing in a wrapped, invalidatable orchestrator, we can prevent `collectPayment` and `retryPayment` calls from doing anything when the  attempt should have been cancelled.