13122 Prevent use of stale readers cancelling
We cache readers found in the StripeCardReaderService. Stripe warn us not to reuse these, and after a successful connection we do discard them.

However, after cancelling a discovery, or facing an error, we _don’t_. This leads to us trying to connect to stale reader objects, which results in a never-ending connection attempt.

This commit clears the cache when we cancel or error in the discovery process.