13123 Prevent use of stale readers after error
Previously, when trying a second connection after an error during reader connection, the last found reader would be immediately presented as an option for the user to connect to.

Unfortunately, this was a stale representation of the reader and would be sent even if the reader in question was turned off.

This was due to the state not being reset after a connection error

Discovery errors already cleared the state of the CardReaderService, but errors during connection do not.

This commit fixes that issue by switching the service to a fault state when the user acknowledges the error.