13162 Check unknown server-side capture errors
Sometimes, we can get an error or non-response when there’s an error capturing the payment on the server.

This can happen when there are poor network conditions. It’s possible for the payment request to be successful and handled by WooPayments, but then the response can be lost.

Previously, this was shown as an error, leading merchants to believe the order wasn’t paid.

This commit improves that by adding a check for whether the order is paid, when it receives one of these errors. If the order is paid, it shows the payment success screen.

If it’s not paid, or we can’t refresh the order, the original error is shown.