Revert populating items from subview
By moving this task back to the dashboard view we resolve the issue introduced here: https://github.com/woocommerce/woocommerce-ios/pull/13188#discussion_r1658059617 when refactoring to split responsibilities.

Currently the products would be populated from the remote every time we un-collapse the view, or when we go back from checkout to edit the order.

By reverting the change and moving this call back to the dashboard we just initialize products once for consistency with Android as well.