Update order with custom amount mock
When order is made with custom amount it is returned with auto-draft status, however when we want to collect payment, the status is updated to pending. We're using this order for custom amount payment so pending status is more appropriate for mocks and makes the order to be persisted in the db as expected.