Animate hiding totals fields after delay when processing payment
When totalsViewModel.isShowingTotalsField turns to false, we need to hide totalsFields after a delay with animation.

I decided to use `onChange` modifier to track totalsViewModel.isShowingTotalsFields value. When it turns to false, I perform animation with a delay and turn a local State var isShowingTotalsFields to false. Otherwise, I immediately turn isShowingTotalsFields to true.