13365 Fix custom payment block on iPadOS16
Using iOS 16, attempts to collect a custom amount via the Payments menu always failed, getting stuck when the user tapped `Collect Payment`.

This was due to mixing two different approaches to `navigationDestination` – using a path, and using isPresented variables. This appears to be unsupported, or at least buggy on iOS 16, it’s unclear which as it’s not documented as a limitation.

The issue manifested as the `presentPaymentMethods` binding not being respected, so it never moved on to let the user choose a payment method, but nor could they back out of the screen either.

To fix it, I’ve removed use of the path and standardised on `isPresented` bindings for both collectPayment and paymentMethod destinations.