13431 Stop preventing animations on TotalsView
Previously we prevented animations before the TotalsView appeared, to ease the visual clutter of the transition.

However, when going back and forwards this prevented the “slide in” from working properly, and the visual clutter can be improved in other ways. It’s better if we just let the animations happen.