13495 Cancel payment immediately on addMore tap
When viewing the checkout screen in POS, we prepare the reader for payment – this means there’s a payment in progress, and the reader is “live” to cards being tapped.

If the merchant taps `Add more` (back) to make a change, we cancel the payment, to give them the opportunity to do that.

Previously, we only did this on disappearance of the totals view, however that doesn’t happen until the transitions are complete.

That meant it was possible for the shopper to tap a card after the merchant tapped back, but before the payment was cancelled. Once the card is tapped, we can’t cancel the payment any more, and this resulted in a broken state for POS.

This commit changes the trigger for cancelling the payment to the tap on `addMore`, while retaining the `onDisappear` version for cases such as exiting POS.