Revert to not use observation
We have found some flakiness in the async tests for the banner when using combine, but this only seems to happen in CI, which makes it very time consuming to debug. We can circle back to use observation for this in the future if needed, for the moment the initial solution works well.