13596 Use NavigationStack to fix Try out TTP flow
On iOS 17, the Try out Tap to Pay on iPhone flow was broken.

After the card payment flow was presented, in the preparing state, the `SimplePaymentsSummary` was dismissed from underneath it, while the merchant tapped their card.

This meant that the `SimplePaymentsSummary` and associated `PaymentMethodsView` were no longer available to handle the ongoing payment flow, and the payment flow was never dismissed.

The `NavigationLink.init(destination:isActive:label:)` used by `LazyNavigationLink` was deprecated with iOS 16, but does still work, generally.

In this case, since the flow was only broken for iOS 17, I opted to move to the new style of navigation, which also requires a `NavigationStack`. This approach fixed the issue, and continues to work on iOS 16.