13596 Avoid use of LazyNavigationLink in payments
To fix Try out TTP, we removed `LazyNavigationLink` use from the flow as it relies on a deprecated form of `NavigationLink`, which caused some issues in iOS 17.

The `SimplePaymentsSummary` is part of that flow, and though it worked with this `LazyNavigationLink`, it seemed better to remove that as well.

`SimplePaymentsSummary` isn’t used anywhere else, any more.