13234 Remove Totals dependency on DashboardVM
The TotalsView is the only place we start new orders from, however, it needs to communicate that to other parts of the POS so they can set their state accordingly.

Previously, we did this by having `TotalsView` call `dashboardViewModel.startNewOrder`, which then called `totalsViewModel.startNewOrder`.

This was the only reason that we had a dependency on the Dashboard from Totals.

By using a publisher to emit events when we need to start a new order, this commit breaks that dependency, so now the TotalsView can talk to its view model, and the view model doesn’t know about the dashboard.

This is preparatory work for using the `startNewOrder` action in the payment capture error actions.