Use TextField instead of BindableTextfield in FormattableAmountTextField
FormattableAmountTextField is mainly used in Order -> Custom Amount flow.

Internally it used a custom BindableTextField.

BindableTextField has issues with focusing text field when the view appears. However, it supports `text` binding not only for notifying on input/output changes, but also allowing to format `text` within viewModel and be reflected in an input field. However, this functionality is not supported with a native TextField.

To avoid BindableTextField issues I'm switching o a native TextField. To support both binding and formatting:
- Removing property observer from ViewModel.amount, since property observers are not meant to work with publishers
- Created a separate updateAmount method from which the updates need to be done
- Created a separate textFieldInput publisher that is used only to reflect input field input but not to be used or accessed outside text field context