13517 Group POS list elements with their title
This ensures VoiceOver steps from the relevant title, through the other elements in the same visual hierarchy.

Without this modifier, it can become unclear as the reader goes in the reading direction (e.g. left to right) across the whole screen, so products in the list and products in the cart are hard to distinguish, and it’s hard to use the headings as navigational tools.