13517 Correctly annouce simple products banner in a11y
In VoiceOver, we don’t want to show the info label as it’s decorative.

We also need to make it clear that the text is all a button (to open “learn more”, which should be styled as a link.)

Finally, we need to make it clear that the x button dismisses the banner, “close” isn’t quite right.