13517 Make it easier to access menu and card reader
This change ensures that all the content is first in the VoiceOver order, and follows the direction of attention within the content (i.e. Product list is primary when building, and the Totals view is primary when finalising)

The floating items follow the content in the accessibility order, as they are a tertiary concern.

Previously, they would generally be read between the first and second content column.