13678 Fix Tap to Pay connection after errors
This commit applies a similar fix to one we previously made for external card readers, in c5827fc2f50992fc85910135d1c070776fb17f6f.

Discovered readers are provided via a `CurrentValueSubject`, and the `StripeCardReaderService` is a singleton so if not managed carefully, old state can persist between connection attempts.

Errors during connection to the card reader (rather than during discovery) don’t set the status of the `StripeCardReaderService` to fault, so the next time you attempt to connect to a reader, it’s in a mid-connection state, and the subject has not been cleared.

This results in the service trying to connect to the old representation of the TTP reader again, using a stale reader object. When we try this, the app becomes unresponsive as Stripe don’t return any error, and the connection process never moves forward. Stripe don’t provide any way to cancel an in-progress connection, so there’s no option to add a timeout to the connection.

With this commit, we switch the status to fault when we get the connection error. This resets the `discoveredReadersSubject`, and prevents the service from return the stale representation of the TTP reader, so we don’t try to connect to it and avoid the unresponsive state. Instead, the next attempt starts from scratch, and can succeed.