13234 Present all POSModals from the root of POS
For the ZStack blurred overlay approach to modal presentation to reliably cover the whole screen, we need to present POSModals from a view which covers the screen.

Moving the presentation logic to `POSRootModalViewModifier` allows us to ensure that the ZStack is always presented with the full size of the POS as the `content` it modifies

Retaining the `posModal` modifier allows us not to care where we present the modals from.