13234 WIP: use item binding for POSModal
Previously, the card reader connection state didn’t update, which left the app in a frozen state, saying it was “scanning for readers” even though it had actually connected to one.

By changing the `posModal` modifier to take binding to an item, the modal gets rebuilt when the alertType for card reader connections changes. This includes the content updates, so we avoid the frozen state.

In this commit, `posModal` only supports the `Binding<Item?>` style of presentation trigger, i.e. when the binding isn’t nil, present the content and pass the content viewBuilder the item.

Existing `isPresented` versions are simply showing when the flag is set to true, so it would be a bit clumsy to start using `item` for those.

This is WIP in that the `isPresented` modals are simply commented out, so we can implement a second version of the `posModal` function in a later commit.