13234 Allow posModals to be controlled via a flag
The `isPresented` version is convenient for simple, unchanging views.

It has some (documented) issues around updates to the content in the viewbuilder – the observation of external objects is broken after the modal is first presented.

It’s provided for convenience only, the `item` version is more reliable for complex views.