13234 Use IDs to compare structs with closures
Closures can’t be compared, which makes it difficult to identify differences between objects which hold them as properties.

Where they can’t be changed, including a pseudo-random ID on an object is a helpful tool to stop us from saying that two buttons with the same title but different actions are equal or have the same hash.

If the closures are held in a `var`, we would need to instead update a random identifier (*not* the object’s ID) every time they were set. The random identifier can then act as a proxy for that closure in comparison operations.

It’s still not perfect, but it’s safer.