13386 Handle reader disconnection on TotalsView
When the TotalsView is showing, we need to handle reader disconnection with some care.

This commit ensures that any in-progress payment is cancelled when the reader disconnects – i.e. when we’ve prepared the reader for payment, but the shopper hasn’t tapped their card, that PaymentIntent should be cancelled before we disconnect.

Additionally, if we connect a reader after disconnecting on the TotalsView, it ends up showing just the Totals, with no next step, because there’s nothing to trigger the payment again. This commit observes for reconnection after we disconnect the reader, and prepares it for payment when we get a connection.