13386 Prevent reader disconnection after card tap
Reader disconnection after the merchant has tapped their card _does work_, but the UI isn’t well designed for it (e.g. the “Connect your reader” showing with a purple or green background.)

In this commit I’ve disabled it when these fullscreen views are shown, as there’s little need to allow it to be disconnected there, and this simplifies the UI state a bit.