13800 Make message to processing transition smooth
The transition when tapping a card would fail, because the payment service emits 3 events in quick succession, too fast for the 0.8s transitions to complete before the next one started.

This makes the transition shorter, so that it succeeds and works smoothly.

There is a more complex approach we can follow up with, to space out the events without losing them, but for now, this will do.