13824 Update Sentry to avoid crash in SDK init
Sentry found a crash in their SDK initialisation process: https://github.com/getsentry/sentry-cocoa/issues/4280

It’s difficult to assess the impact, as we don’t have the benefit of any Sentry issues for these crashes. We believe at least 12 devices have been affected in the last 2 weeks.

This commit updates us to the latest version, which has the fix for the crash.