13829 Use unformatted total to calculate change
Passing a formatted currency through `CurrencyFormatter.convertToDecimal` can easily fail, as we don’t know enough about how it’s formatted to reliably unformat it, which we need to do to convert it to a decimal.

This change passes around the unformatted amount more as well, to avoid issues calculating change, for example:

- on amounts greater than 1000, e.g. $1,000.00 we previously interpreted both `,` and `.` as decimals, and then converted the number to $1
- for currencies with other separators, e.g. space, we would also trim at the first space
- removing currency symbols is prone to fail if using settings like `left with space`

`convertToDecimal` is still flawed, but works OK-ish for changing API amount strings to a decimal. These API strings reliably have no thousand separator or currency symbol, and always use a `.` as the decimal separator. The number of decimals changes according to the WooCommerce settings.

Note that I’ve marked SimplePayments as deprecated. Currently, this is only used for `Try out Tap to Pay on iPhone`, and it should stay that way. I wasn’t reasonably able to get the unformatted total for simple payments in a resilliant way, as the logic is a little complex – however, for the very simple and geographically contained use of TTP, this is fine for now. I just don’t want further use of Simple Payments to creep in.