13829 Use unformatted amount to calculate discount
We were repeatedly passing amounts through `CurrencyFormatter.convertToDecimal` and `formatAmount` when setting discounts.

This added risk with different locale formats that we would make a mistake converting the string, and end up with a discount of over 1000 having anything after the thousands separator truncated, i.e. a discount of $1,000.00 was applied as $1.

This commit reduces the number of conversions we do, by only converting for display, and using the decimal input for calculations.