Adds animation to hide TabBar when entering POS
This required a bit of trial and error: Depending on how we call the animation and its completion handler, there’s a risk of a regression to gh-13785 where the UI is not being updated appropiately, leaving additional un-used space at the bottom of the screen. Please check the comment in the function for more context.