13879 CardPaymentService: Vend connection status
As preparation for immediately dismissing reader connection modals, we need to provide more granular detail about the current state of the connection process from the CardPresentPaymentService.

The first step of that is to make it responsible for the `disconnecting` state, so that we can provide the similar `connectionCancelled` state when a user cancels the connection.

The goal is to disable the button while the cancellation of the connection continues, even after the modal has been dismissed.