Link Yosemite to NotificationExtension target
The changes in 1fd27830a602886f96f5bfe2b6c346e851880041 and
2e7b7e063c4d8e5446dc7d5a6962f1405ff6d2a0 add `import Yosemite` to
`MarkOrderAsReadUseCase.swift`.

That file is part of WooCommerce, WooCommerce Watch App, and
NotificationExtension.

WooCommerce links to Yosemite. The Watch target does not, but it's not a
problem because of the `#if canImport(Yosemite)` check.

But what about NotificationExtension? It did not have Yosemite before
this change. Might it be the reason for the CI failure:

```
ld: warning: Could not find or use auto-linked framework 'Yosemite': framework 'Yosemite' not found
❌ Undefined symbols for architecture arm64
❌   "protocol witness table for Yosemite.NotificationAction : Yosemite.Action in Yosemite", referenced from:
❌   "nominal type descriptor for Yosemite.NotificationAction", referenced from:
❌   "type metadata for Yosemite.NotificationAction", referenced from:
❌ ld: symbol(s) not found for architecture arm64
❌ clang: error: linker command failed with exit code 1 (use -v to see invocation)

[...]

The following build commands failed:
	Ld /opt/ci/builds/builder/automattic/woocommerce-ios/DerivedData/Build/Products/Debug-iphonesimulator/NotificationExtension.appex/NotificationExtension normal (in target 'NotificationExtension' from project 'WooCommerce')
```

See for example
https://buildkite.com/automattic/woocommerce-ios/builds/24623#0191e996-5541-4a94-9454-30f771bae463/7218-13610