Call keyboard dismissal on tap gesture
We’ve seen some instances where the CTA button is not tappable unless the keyboard is manually dismissed, this issue hasn’t been replicable consistently thought.

We attempt to mitigate this problem by allowing keyboard dismissal when a merchant taps anywhere else within the view. Adding this within the button closure wouldn’t work, as if the button happens to be non-tappable due to the keyboard layout constraints, we wouldn’t be able to call it’s dismissal either.