13879 Dismiss modals without waiting for cancellation
The reader connection modals were previously dismissed by `idle` being sent to the payment events stream when the cancellation action completed.

This meant that the `x` button in the ui would appear to take a second or two to dismiss the modal.

Now that we show `Please wait` while there’s a cancellation in progress, there’s less risk of trying to connect to a reader while a cancellation’s ongoing. That means we can dismiss the connection modal immediately, and allow the merchant to do other things.