Dismiss Onboarding when successful
Onboarding can be either pushed (on iPhone from Collect Payment in settings) or modally presented (in all other contexts.)

When it was pushed onto a nav stack, it would be dismissed automatically when it completed… but when it was modally presented, it would remain on screen until the user swiped it away.

To improve this, we can detect whether the `onboardingViewController` is in a navstack, or modally presented, and react accordingly when dismissing it.

Now, the onboarding screen is correctly removed and payment continues, when it’s completed.