Don’t repeatedly try to hide onboarding
When onboarding has been completed, we hide onboarding and call a completion handler to indicate to the payment systems that we’re ready to collect a payment.

This is done in a Combine subscription handler, but that was not cleared after we called completion.

This adds risk that we might try to hide onboarding twice (which would be fine) and then start the payment twice (which wouldn’t be great!)

This commit clears the subscription when we’ve completed onboarding.