Continue onboarding after selecting payment gateway
Previously, if an IPP onboarding run performed during the payment flow showed the Select Payment Gateway screen, the onboarding would not continue after the merchant selected a plugin. Instead, they would have to tap `Card reader` again, and then see the next onboarding issue.

This commit resolves that by removing the `dismiss` call when the button is tapped, and leaving it to the view model to determine whether the select payment gateway view should be presented or not.