Update error message for Tap to Pay support
The minimum requirements for TTP support have increased to iOS 16.7.

This changes the error message shown when using an older version to tell merchants about the need to update.

We still show the old requirements elsewhere in the app with this change, but this avoids hiding TTP from merchants who’ve previously used it, without giving them any explanation for why it’s gone.