14333 Show IPP address errors with site credentials
Previously, we decoded specific `DotcomError` instances to show the specific errors needed when a user hasn’t set their site address or postcode in WP-Admin, which prevents IPP reader connections from working.

These errors weren’t decoded when using site credentials to log in, as in that case a `NetworkError` is received, and our error handling didn’t decode that correctly.

This adds correct decoding and tests for the `NetworkError` variants of these issues.