Fix StatsStoreV4 deallocation in MockStatsActionV4Handler
When the StatsStoreV4 instance was being created as a local variable in retrieveStats(), etc, it caused it  to be deallocated before the async CoreData  operation could complete. This resulted in stats data not being saved
due to [weak self] returning early in `StatsStoreV4.upsertStoredOrderStats`.

This commit makes StatsStoreV4 a property of MockStatsActionV4Handler to ensure it remains alive throughout async operations. In effect, this allows the Performance card in the screenshot test to show correct order data.