14304 Non-default currency order edit notice
As a workaround, we disable editing orders with a non-default currency. If we allowed this, items on these orders would be shown in the wrong currency. See peaMlT-XF-p2 for context.

Multi-currency is not supported in the app at present, but when we add support properly, this workaround should be removed and the edit screen fixed.