14357 Use store currency for all new orders
On multicurrency sites, the user’s default currency is used for order creation when none is specified in the API request.

This may differ from the site’s default currency. A user can set their default currency at `/my-account/edit-account/`.

Previously we didn’t specify a currency, so if the merchant’s default was different from the site’s, they weren’t be able to use in-person payments.

Additionally, order creation previously showed that they were creating the order in the site’s default currency, but when it was created it actually used the user’s, so it appeared to switch over, which was confusing.

To work around this in a world where the app doesn’t really support multicurrency, we set the site’s default currency code in the initial order creation request.