14350 Use ItemListState to populate item list
This commit removes use of the published `items` array from the ItemListView. It’s still in use on the dashboard – a future ticket will attempt to remove the need for it when creating the order.

One change improves the `initialLoading` logic to only rely on the state, not the item list. That caused pull-to-refresh to break, because it used `items.isEmpty` as an extra gate on using the full-screen loading indicator. A change to the pull to refresh call, to stop using `loadInitialItems`, fixed this as well.

With these changes, we no longer need the itemsPublisher, so it’s removed.