14458 Move reader reconnection to aggregate model
We observe reader reconnection so that we can prepare the card reader for payment as soon as it connects, or reconnects, while on the totals page.

This was being bounced through the aggregate model, dashboard, totals view model, and then back to the aggregate model… which was quite a roundabout way of observing whether we were in the building or finalizing steps.

This moves it to be internal to the aggregate model