Interact with the app before taking the first screenshot
We do this by using `TabNavComponent().goToMyStoreScreen()` instead of directly referencing `MyStoreScreen()`.

addUIInterruptionMonitor's callback is not called until we interact with the app, in the past, we used to use `app.tap()` for this, but since it was removed, the permission dialog would be still visible on the first screenshot. When I attempted re-introducing it back, I noticed that it caused issues on tablets by opening some unwanted screens, so I used `TabNavComponent().goToMyStoreScreen()`, it plays the same role now, as it also allows to interact with the app, and also solves the issue on tablets.