14178 Set custom amount sheet stops by text size
These stops are determined by trial and error based on an iPhone SE. They transpose fairly well to larger phones when using smaller sizes, but optimising for accessibility sizes on the small phone leads to full-screen sheets too quickly on large phones.

Since users who want very large text are likely to choose a larger screen, I’ve opted for giving the choice between medium and large at these larger sizes, so they can still see the context of the order if they choose. On an SE, there will be some cropping in these sizes, but it can be scrolled.