Kind-of working variation display
This uses a flat ItemListState, which could show root or child sets of items.

There’s an opportunity to use this in a NavigationStack, however we can’t do that while POS is presented _within_ a navigation stack. That’s a bit of a roadblock to properly animating changes, unless we reimplement it :grimace:

There are issues here (hence “kind-of”). Sometimes when opening a variation, especially just after loading or refreshing the main list, it sometimes shows the main list again not the variations. I’ve not explored why, but I suspect there’s some race condition :/