14699 Avoid changing the view state during refresh
Previously we sent a `.loading` state in the pull to refresh handler.

That changes the POSModel observable object, which results in the whole POS being redrawn.

Redrawing the refresh controller cancels the ongoing async task so it always shows an error.

With this change, we lose the skeleton loading cell when pulling to refresh, but the function works and the default control still shows.

With more granular state observation, we could update the list content without redrawing the list component itself, avoiding this issue.