Remove shared Workspace `xcconfig` because it was not shared
It turned out that every project made a copy of that file, which
defeated the whole purpose.

Apparently, we cannot have files shared at the workspace level.

It might still be possible to achieve the same result in the future, but
that would require every project being set up with base configuration files
and referencing the shared file from there, without adding the shared file
itself to the projects. But to do that, we'll need each project set up
that way, which is out of scope for the work I'm currently doing of
folding the WordPressAuthenticator and WordPressKit libraries into
WooCommerce iOS.