isCashPaymentInProgress flag to intercept card events
Each time we tap the card on the reader, the .card payment state would override the current .cash payent state, changing the views. With this flag we filter card events when a cash transaction is in progress, so we can ignore them.

Allowing card payments while cash hasn’t been “marked as complete” yet, and disallowing doubles charges on the success screen.