14696 Improve pagination trigger
Previously we tracked the scroll position and compared it to a threshold based on view height. This resulted in loading the next page whenever the view is scrolled.

This changes the approach to show the loading cell whenever we get to the bottom of the list, as long as there are still pages to load.

When the loading cell appears, which with the move to LazyVStack is only when it’s scrolled in to view, we start the next page fetch request.