14357 Use initial order to determine changes
The order synchronizer previously used a static let of an empty order to make it’s new order.

The same static property was used to check whether the order had been edited or not.

Because Order has dates in it, the static property would have had the dates from the initial call to it, as long as the app was running.

When we added the currency parameter, this made it clunkier to use the static property approach, and it made little sense to have out of date dates anyway.

This commit updates the order sychronizer to be responsible for keeping track of whether the order has changed or not. It’s still done using equality, but we compare to the order that we initially created, rather than the static property.