14905 Show a full page of ghost cells when loading
Previously, we showed a single ghost cell when loading the next page, or the first page of variations.

The most common case is that more than 1 item will load with each page; most likely the same number as the page size request.

Showing the same number of ghost cells as will show on the screen after the load is complete is the ideal situation

We can’t show 25 rows on screen at a time anyway, so this commit increases the number of ghost cells to 8. This is enough to fill the portrait screen on the smallest font size.

Now, we show a skeleton layout which more closely matches the eventual loaded layout. This approach matches Android’s behaviour too.