14834 Only apply new WC version req when flag true
To retain the ability to turn off the feature flag for variable products, we should only use the higher minimum WC version when the flag is enabled.

If we disable the flag, downloadable products are still filtered out by local filters, and it’s no longer necessary to remove virtual products.