14409 Handle capture errors when using site creds
When using site credentials, we get `NetworkError` not `DotcomError`.

Previously, we handled the `orderCaptureError` as a special case when we got it while logged in via Jetpack, but as a generic error when using site credentials

By using the new `PaymentError`, we can decode the `NetworkError` responses for `orderCaptureError` and use the special case handling however you log in.