13778 Improve Try Again for card reader updates
Previously, card reader updates which failed due to low battery had a single button, `Cancel`.

However, this opened the `Multiple readers found` screen without ever showing any more readers. This was intended as a retry function, so that you could plug in the reader and then connect and redo the update.

The `multiple readers found` screen never found any more readers because the `StripeCardReaderService` had already been reset because of the error.

A better solution is to provide both a `Cancel` and `Try again` button, and have the retry actually start the connection process from scratch.

Previously we did this using `.beginSearch`, however this doesn’t reset the `CardReaderConnectionController`’s state, which could lead to a reader being ignored incorrectly, or the multiple readers screen be shown with zero or one reader on it. Using `initializing` lets us truly start from scratch here.