14992 Publish all reader status changes
When we get a reader status change from the underlying `CardPresentPaymentStore`, we should always republish that to POS.

Previously, we would treat empty arrays of readers from the store (nil in the `connectedReaderPublisher`) as something we don’t have to pass on, when in reality they likely represent a disconnection originating outside the app (e.g. turning the reader off, bluetooth failure, or the battery running out.)

This commit updates the handling to treat these cases as a `disconnected` reader, so now the UI will update accordingly.