13738 Don’t show Connect Your Reader on success
Previously, it was possible to see the `Connect your reader` message in inappropriate parts of the flow, where it could cause confusion.

For example, if we were on the Payment Successful, or other payment related views, and lost connection to the card reader (from battery drain, switched off, bluetooth disabled,) then it would show `Connect your reader`.

After connecting, the POS would be in a strange state, and it could lead the merchant to thinking a new order was paid for when actually it was the previous order.

This commit prevents the disconnected view from showing unless the payment is still to be taken.