Fix back from cash screen for free orders
Previously, when the card reader was connected and the merchant tapped `Back` on the cash collection screen for a free order, nothing happened.

This is because we rely on `collectCardPayment` to update the payment state to `.card`.

By setting it to idle first, we ensure we always go back, then we might try taking a payment.