13251 Prevent POS use in horizontally compact view
There’s not enough space for our two-pane layout to work when you get to a horizontally compact display, so until we add support for compact layouts, e.g. on the phone, we’ll show an overlay advising the merchant to give it a bit more space.

This still allows use in landscape 2/3 view on all iPads, and 1/2 view on 12-13” iPads. The 1/2 view can be improved by hiding images on cart items under a certain screen width; I’ll do that separately.