14869 Move refresh control to fix refresh issue
Now only the ItemListView and ChildItemList are redrawn when we pull to refresh, not the dashboard.

This means that we can avoid redrawing the `refreshable` modifier by moving it up the heirarchy, to where the views are used.

Moving it up means that we can continue to use the state in the view itself, without cancelling the refresh task as soon as it starts, and consequently showing an error.

This is only needed because we change the list state when we pull to refresh, to `.loading`, which should arguably be reserved for loading the next page (as it adds the ghost cell at the end of the list)