14869 Hide cancellation errors when refreshing
If we change the view content while the refresh is ongoing, e.g. by loading a new page, the refresh request will be cancelled.

It doesn’t make sense to show that as an error to the user.

Instead, we swallow that error, and set the state back to `loaded` with all the items that we’ve got up to now.